% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QExplorer.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning the data set}
\usage{
eq_clean_data(df)
}
\value{
A location name that has a properly formatted text for analysis and graphing, latitude and longditude are numeric and the date is properly formatted.
}
\description{
This function cleans the original raw data set from the NOAA Signficant Earthquake Database and formats some columns to be used in exploratory analysis.
}
\note{
This function cleans key fields of the earth quake data set for exploratory analysis.
}
