% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_explorer.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function for labeling spatial analysis of earthquake data (earth mapping)}
\usage{
eq_create_label(data)

eq_create_label(data)
}
\description{
This function provides a spatial interpretation of earthquake data with popups of location, magnitude and mortality.

This function provides a spatial interpretation of earthquake data with popups of location, magnitude and mortality.
}
\note{
For BC date ranges, use a one sided inequality e.g. datevalue < '0000-01-01'

For BC date ranges, use a one sided inequality e.g. datevalue < '0000-01-01'
}
\examples{
\dontrun{

# User inputs a country and range of dates to be explored.
# The earth map will plot the points as a leaflet with a pop up based on the date of the event.
# For BC date ranges, use a one sided inequality e.g. datevalue < '0000-01-01'

library(leaflet)
library(dplyr)

any_name_df<-eq_clean_data(results)

any_name_df<-any_name_df\%>\%filter(COUNTRY=="GREECE")\%>\% filter(datevalue > '1900-01-01', datevalue < '1950-01-01')
any_name_df<-any_name_df\%>\%mutate(popup_info=eq_create_label(any_name_df))
any_name_df\%>\%leaflet() \%>\% leaflet::addTiles()\%>\%addCircleMarkers(lng = ~LONGITUDE,lat = ~LATITUDE,radius = ~EQ_PRIMARY,weight=1,popup=~popup_info)

}

\dontrun{

# User inputs a country and range of dates to be explored.
# The earth map will plot the points as a leaflet with a pop up based on the date of the event.
# For BC date ranges, use a one sided inequality e.g. datevalue < '0000-01-01'
# Not all events will have a mortality value.

library(leaflet)
library(dplyr)

any_name_df<-eq_clean_data(results)


any_name_df<-any_name_df\%>\%filter(COUNTRY=="GREECE")\%>\% filter(datevalue > '1900-01-01', datevalue < '1950-01-01')
any_name_df<-any_name_df\%>\%mutate(popup_info=eq_create_label(any_name_df))
any_name_df\%>\%leaflet() \%>\% leaflet::addTiles()\%>\%addCircleMarkers(lng = ~LONGITUDE,lat = ~LATITUDE,radius = ~EQ_PRIMARY,weight=1,popup=~popup_info)

}
}
