% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QExplorer.R
\name{import_quake_data}
\alias{import_quake_data}
\title{Reading an earth quake datadata set}
\usage{
import_quake_data(filename)
}
\arguments{
\item{filename}{Filename in csv format to be read. The file must be located in the working directory.}
}
\value{
Returns a data frame from a csv file into tabular format.
}
\description{
This function reads in an exported data set from the NOAA Signficant Earthquake Database.
}
\note{
Data must have a .csv extension.  The NOAA data is exported as a tab delimited file type without a file extension designation.
}
\references{

}
