% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QExplorer.R
\name{simpleCap}
\alias{simpleCap}
\title{Simple capitalization routine}
\usage{
simpleCap(x)
}
\arguments{
\item{Input}{A generic string.}
}
\value{
A character vector in which the first letter is capitialized of each string segment.
}
\description{
This simple function capitalizes the first letter of a string (if it's in lower case).
}
